Refactor dashboard component
Merge from master
