+++ src/screens/General/Dashboard/index.js
@@ -1,0 +1,4 @@
+export { default as EventsList } from "./EventsList";
+export { default as FavoriteCommittees } from "./FavoriteCommittees";
+export { default as Leaderboard } from "./Leaderboard";
+export { default as Dashboard } from "./Dashboard";
